Wallet restructure + add fees
This restructure changes the way lightsats stores tip and user balances.
Rather than having a single wallet, each unclaimed tip has a wallet.
Each user has a staging wallet.
Claiming/Reclaiming tips will move the amount+fee from the tip wallet into the user's staging wallet.
When the user withdraws, only the original tip amount will go to the user.
The fee will be used to pay the outbound routing fee.
If there is any remainder of the fee left over, it will be transfered to the lightsats wallet (used for infrastructure/dev costs)

+++ app/pages/api/tipper/tips/[id].ts
@@ -45,7 +45,41 @@ async function deleteTip(
   req: NextApiRequest,
   res: NextApiResponse<Tip>
 ) {
+  if (!process.env.LNBITS_API_KEY) {
+    throw new Error("No LNBITS_API_KEY provided");
+  }
+
   if (tip.status === TipStatus.UNFUNDED) {
+    const lnbitsWallet = await prisma.lnbitsWallet.findUnique({
+      where: {
+        tipId: tip.id,
+      },
+    });
+    if (lnbitsWallet) {
+      const deleteWalletRequestHeaders = new Headers();
+      deleteWalletRequestHeaders.append(
+        "X-Api-Key",
+        process.env.LNBITS_API_KEY
+      );
+
+      const deleteLnbitsUserResponse = await fetch(
+        `${process.env.LNBITS_URL}/usermanager/api/v1/users/${lnbitsWallet.lnbitsUserId}`,
+        {
+          method: "DELETE",
+          headers: deleteWalletRequestHeaders,
+        }
+      );
+      console.log(
+        "Tip",
+        tip.id,
+        "Delete tip lnbits user + wallet response: ",
+        deleteLnbitsUserResponse.status,
+        deleteLnbitsUserResponse.statusText
+      );
+    } else {
+      console.warn("No lnbits user+wallet for tip " + tip.id);
+    }
+
     await prisma.tip.delete({
       where: {
         id: tip.id,
